Revert "Clients can connect to witness"
This reverts commit 63722962a17225e8fce363a27200bc3d8a3ef42a.
