+++ be1-go/cli/pop.go
@@ -71,13 +71,7 @@ func main() {
 								Name:    "organizer-port",
 								Aliases: []string{"op"},
 								Usage:   "port on which to connect to organizer websocket",
-								Value:   9001,
-							},
-							&cli.IntFlag{
-								Name:    "client-port",
-								Aliases: []string{"cp"},
-								Usage:   "port to listen websocket connections from clients on",
-								Value:   9002,
+								Value:   9000,
 							},
 						},
 						Action: witness.Serve,
